requirements.txt
appdirs==1.4.4
Appium-Python-Client==2.6.0
apturl==0.5.2
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==2.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
black==21.12b0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
fasteners==0.14.1
filelock==3.0.12
fonttools==4.38.0
future==0.18.2
h11==0.13.0
httplib2==0.14.0
idna==2.8
iniconfig==1.1.1
ipython==8.0.1
jedi==0.18.1
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
lark-parser==0.12.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.6.2
matplotlib-inline==0.1.3
monotonic==1.5
mypy-extensions==0.4.3
netifaces==0.10.4
networkx==2.8.8
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.26
protobuf==3.6.1
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pycairo==1.16.2
pycodestyle==2.9.1
pycups==1.9.73
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.6
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.3.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stack-data==0.1.4
systemd-python==234
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.0.1
tzdata==2022.6
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.10
usb-creator==0.3.7
virtualenv==20.4.6
wadllib==1.3.3
wcwidth==0.2.5
wsproto==1.1.0
xkit==0.0.0
youtube-dl==2020.3.24
